chore: separate test command from root
